update recastnavigation
https://github.com/cmangos/mangos-classic/commit/9b38ef751b83cd091267140153a1c383b7c769c4#diff-2d62c0f4571f022165bb703dbe4cb7b0

+++ src/game/Maps/MoveMap.h
@@ -32,7 +32,7 @@
 #include "Detour/Include/DetourNavMeshQuery.h"
 
 //  memory management
-inline void* dtCustomAlloc(int size, dtAllocHint /*hint*/)
+inline void* dtCustomAlloc(size_t size, dtAllocHint /*hint*/)
 {
     return (void*)new unsigned char[size];
 }
